Add actual output of the command (read more)
I don't know if this is the desired effect, but this the correct output
I get when running the example code.

Personally it looks unnecessary and for me it's enough to set the
`.scriptName` to "", but I'm adding it for now since that's what the
command outputs. You might want to remove this behaviour and then later
update this commit.

Signed-off-by: Kipras Melnikovas <sarpixo@gmail.com>