Remove bad trailing spaces (read more)
Git doesn't show it, but in `index.html` from line 62 to 77 there were
additional spaces added. When you copy this code from the docs, it
doesn't detect the shebang (`#!/usr/bin/env node`) correctly and adds
unnecessary spaces which - if you don't have a formatter - will be
annoying to remove.

I know that it might do with the visuals of the docs, but you could
easily just add some padding for the code blocks instead and not
sacrifice the experience of developers who just want to quickly try out
the sample code:)

Signed-off-by: Kipras Melnikovas <sarpixo@gmail.com>